[Feature] Merge upstream 105.0.5144.1

+++ base/win/patch_util.h
@@ -16,10 +16,12 @@ namespace internal {
 // Copies |length| bytes from |source| to |destination|, temporarily setting
 // |destination| to writable. Returns a Windows error code or NO_ERROR if
 // successful.
-BASE_EXPORT DWORD ModifyCode(void* destination, const void* source, int length);
+BASE_EXPORT DWORD ModifyCode(void* destination,
+                             const void* source,
+                             size_t length);
 
 }  // namespace internal
 }  // namespace win
-}  // namespace bsae
+}  // namespace base
 
 #endif  // BASE_WIN_PATCH_UTIL_H_
